# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ms-CRM OpenApi Documentation",
        "description": "ms-CRM OpenApi description",
        "contact": {
            "email": "florian@alt.bzh"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/",
            "description": "Full ms-CRM API endPoint documentation"
        }
    ],
    "paths": {
        "/company/{id}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "get company by id",
                "description": "Afficher la fiche d'une société",
                "operationId": "338664ac2f96ec0d3a64676c70a4aeb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return company object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "no company found"
                    }
                }
            },
            "put": {
                "tags": [
                    "company"
                ],
                "summary": "Update company",
                "description": "Mise à jour de la fiche d'une société",
                "operationId": "ef091f3b4f281d3762a0283d19b5ec0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id's company to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return company object updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "company"
                ],
                "summary": "Delete company",
                "description": "Archivage de la fiche d'une société",
                "operationId": "963a45f9ab5bbef4f7601d48fa76e74d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id's company to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return company object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Company not deleted"
                    }
                }
            }
        },
        "/company/withUser/{id}": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "List of all users for one company",
                "description": "Affichage de tous les utilisateurs reliés à cette compagnie",
                "operationId": "9a6fa1eecd544ddbe49ea1b7900daeb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return company object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No company store"
                    }
                }
            }
        },
        "/company/list": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "Get list of all companies",
                "description": "Liste de toutes les fiches de sociétés",
                "operationId": "1e3d5919745b50fcd4d9ff8d784b239a",
                "responses": {
                    "200": {
                        "description": "return company object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No company's list found"
                    }
                }
            }
        },
        "/company": {
            "post": {
                "tags": [
                    "company"
                ],
                "summary": "Store company",
                "description": "Création de la fiche d'une société",
                "operationId": "e4ed87f74ea810e8c76674f6203cc09f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return company object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No company store"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CompanyStoreRequest": {
                "title": "CompanyStoreRequest",
                "description": "Company request body data",
                "required": [
                    "name",
                    "status_id"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of company",
                        "type": "string",
                        "example": "Welch Fargo Company"
                    },
                    "siret": {
                        "title": "siret",
                        "description": "siret'number's company",
                        "type": "integer",
                        "example": "29836108"
                    },
                    "vat_number": {
                        "title": "vat_number",
                        "description": "vat_number's company, for invoices",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "vat_applicable": {
                        "title": "vat_applicable",
                        "description": "vat_number's company, for invoices",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "type": {
                        "title": "type",
                        "description": "type's company, refer to customfield's db",
                        "type": "integer",
                        "example": "2"
                    },
                    "activity": {
                        "title": "activity",
                        "description": "activity's company, refer to customfield's db",
                        "type": "integer",
                        "example": "34"
                    },
                    "address": {
                        "title": "address",
                        "description": "address' company",
                        "type": "string",
                        "example": "23, rue de la liberté"
                    },
                    "zip": {
                        "title": "zip",
                        "description": "zip of city",
                        "type": "integer",
                        "example": "49100"
                    },
                    "city": {
                        "title": "city",
                        "description": "city's address of company",
                        "type": "string",
                        "example": "Angers"
                    },
                    "social_network": {
                        "title": "social_network",
                        "description": "social_network's company",
                        "type": "string",
                        "example": "je sais pas trop"
                    },
                    "status_id": {
                        "title": "status_id",
                        "description": "status_id of company, refer to customfield's db",
                        "type": "integer",
                        "example": "7"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "id of user referent for this company, if it's already known",
                        "type": "integer",
                        "example": "99"
                    },
                    "firstname": {
                        "title": "firstname",
                        "description": "firstname of user referent for this company, if it's a new user",
                        "type": "string",
                        "example": "Amélie"
                    },
                    "lastname": {
                        "title": "lastname",
                        "description": "lastname of user referent for this company, if it's a new user",
                        "type": "string",
                        "example": "Poulain"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user referent for this company, if it's a new user",
                        "type": "string",
                        "example": "brocolidu49@outlook.ru"
                    },
                    "role": {
                        "title": "role",
                        "description": "role of user referent for this company, refer to customfield's db",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "CompanyUpdateRequest": {
                "title": "CompanyUpdateRequest",
                "description": "Company update request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of company",
                        "type": "string",
                        "example": "Welch Fargo Company"
                    },
                    "siret": {
                        "title": "siret",
                        "description": "siret'number's company",
                        "type": "integer",
                        "example": "29836108"
                    },
                    "vat_number": {
                        "title": "vat_number",
                        "description": "vat_number's company, for invoices",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "vat_applicable": {
                        "title": "vat_applicable",
                        "description": "vat_number's company, for invoices",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "type": {
                        "title": "type",
                        "description": "type's company, refer to customfield's db",
                        "type": "integer",
                        "example": "2"
                    },
                    "activity": {
                        "title": "activity",
                        "description": "activity's company, refer to customfield's db",
                        "type": "integer",
                        "example": "34"
                    },
                    "address": {
                        "title": "address",
                        "description": "address' company",
                        "type": "string",
                        "example": "23, rue de la liberté"
                    },
                    "zip": {
                        "title": "zip",
                        "description": "zip of city",
                        "type": "integer",
                        "example": "49100"
                    },
                    "city": {
                        "title": "city",
                        "description": "city's address of company",
                        "type": "string",
                        "example": "Angers"
                    },
                    "social_network": {
                        "title": "social_network",
                        "description": "social_network's company",
                        "type": "string",
                        "example": "je sais pas trop"
                    },
                    "status_id": {
                        "title": "status_id",
                        "description": "status_id, refer to customfield's db",
                        "type": "integer",
                        "example": "7"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "id of user referent for this company, if it's already known",
                        "type": "integer",
                        "example": "99"
                    },
                    "firstname": {
                        "title": "firstname",
                        "description": "firstname of user referent for this company, if it's a new user",
                        "type": "string",
                        "example": "Amélie"
                    },
                    "lastname": {
                        "title": "lastname",
                        "description": "lastname of user referent for this company, if it's a new user",
                        "type": "string",
                        "example": "Poulain"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user referent for this company, if it's a new user",
                        "type": "string",
                        "example": "brocolidu49@outlook.ru"
                    },
                    "role": {
                        "title": "role",
                        "description": "role of user referent for this company, refer to customfield's db",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "Company": {
                "title": "Company",
                "description": "Company model",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name of company",
                        "type": "string",
                        "example": "Welch Fargo Companie"
                    },
                    "siret": {
                        "title": "siret",
                        "description": "siret'number's company",
                        "type": "integer",
                        "example": "29836108"
                    },
                    "vat_number": {
                        "title": "vat_number",
                        "description": "vat_number's company",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "vat_applicable": {
                        "title": "vat_applicable",
                        "description": "vat_number's company",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "type": {
                        "title": "type",
                        "description": "type's company, refer to customfield's db",
                        "type": "integer",
                        "example": "2"
                    },
                    "activity": {
                        "title": "activity",
                        "description": "activity's company, refer to customfield's db",
                        "type": "integer",
                        "example": "34"
                    },
                    "address": {
                        "title": "address",
                        "description": "address' company",
                        "type": "string",
                        "example": "23, rue de la liberté"
                    },
                    "zip": {
                        "title": "zip",
                        "description": "zip of city",
                        "type": "integer",
                        "example": "49100"
                    },
                    "city": {
                        "title": "city",
                        "description": "city's address of company",
                        "type": "string",
                        "example": "Angers"
                    },
                    "social_network": {
                        "title": "social_network",
                        "description": "social_network's company",
                        "type": "string",
                        "example": "je sais pas trop"
                    },
                    "status_id": {
                        "title": "status_id",
                        "description": "status_id, refer to customfield's db",
                        "type": "integer",
                        "example": "7"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Company"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}